nn-model/requirements.txt
# Requirements for κ Predictor Neural Network

# Core ML/DL libraries
torch>=2.0.0
scikit-learn>=1.3.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Optional: GPU support (uncomment if using CUDA)
# torch-cuda>=2.0.0
